extract stream_base
