+++ src/lib/asio/info.txt
@@ -9,6 +9,7 @@ asio_async_handshake_op.h
 asio_async_read_op.h
 asio_async_write_op.h
 asio_convert_exceptions.h
+asio_stream_base.h
 asio_stream_core.h
 asio_includes.h
 </header:public>
